Rework to for CodSpeed
See https://codspeed.io/advent

I'm intrigued to see what good looks like for this thing. I'm unlikely
to be up early enough to submit answers that get on the AoC leaderboard
but maybe I can write something performant enough to get on a different
leaderboard? Things I've written in leetcode and CMU 15-445/645 have
been OK in the past...

So to make that happen, I think I need to expose `part1` and `part2`
functions with appropriate signatures as a public module on the crate.
Let's see.